+++ src/main/java/com/example/jobsearch/model/RespondedApplication.java
@@ -2,7 +2,7 @@ package com.example.jobsearch.model;
 
 public class RespondedApplication {
     private Long id;
-    private Integer resumeId;
-    private Integer vacancyId;
+    private Long resumeId;
+    private Long vacancyId;
     private Boolean confirmation;
 }
